Food

+++ Assets/Scripts/GlobalConfig.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 [CreateAssetMenu(fileName = "GlobalConfig", menuName = "Scriptable Menu/Global Menu", order = 1)]
 public class GlobalConfig : ScriptableObject {
-    public int GridSizeAbsolute;
+    public float GridSizeAbsolute;
     public int GridColumns;
-    public int CellSize;
+    public float CellSize;
 }
